Extract main javascript from its UMD wrapper
